feat: implementing redis DB

+++ cmd/api/src/controller/auth/dto.rs
@@ -9,5 +9,5 @@ pub struct PostulanteLoginRequestDTO {
 #[derive(Serialize, Deserialize)]
 pub struct PostulanteLoginResponseDTO {
     pub token: String,
-    pub expires_in: u32,
+    pub expires_in: u64,
 }
